Hoppas det går

+++ backend/src/main/java/com/example/journalsystem/Repository/UserRepository.java
@@ -12,6 +12,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     List<User> getUserByEmail(String email);
 
-    Optional<User> getUserById(int id);
+    Optional<User> getUserById(Long id);
 
 }
